+++ test/e2e/util.go
@@ -5,18 +5,31 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"log"
 	"net/http"
 	"testing"
 	"time"
 
 	paasapi "github.com/belastingdienst/opr-paas/v3/api"
 	"github.com/belastingdienst/opr-paas/v3/api/plugin"
+	"github.com/belastingdienst/opr-paas/v3/api/v1alpha1"
+	"github.com/belastingdienst/opr-paas/v3/api/v1alpha2"
 	"github.com/belastingdienst/opr-paas/v3/pkg/fields"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/portforward"
+	"k8s.io/client-go/transport/spdy"
 
+	quotav1 "github.com/openshift/api/quota/v1"
+	userv1 "github.com/openshift/api/user/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/runtime"
 	apimachinerywait "k8s.io/apimachinery/pkg/util/wait"
 	"sigs.k8s.io/e2e-framework/klient/k8s"
+	"sigs.k8s.io/e2e-framework/klient/k8s/resources"
 	"sigs.k8s.io/e2e-framework/klient/wait"
 	"sigs.k8s.io/e2e-framework/klient/wait/conditions"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
@@ -35,10 +48,10 @@ var (
 
 // deleteResourceSync requests resource deletion and returns once k8s has successfully deleted it.
 func deleteResourceSync(ctx context.Context, cfg *envconf.Config, obj k8s.Object) error {
-	resources := cfg.Client().Resources()
-	waitCond := conditions.New(resources).ResourceDeleted(obj)
+	cliResources := cfg.Client().Resources()
+	waitCond := conditions.New(cliResources).ResourceDeleted(obj)
 
-	if err := resources.Delete(ctx, obj); err != nil {
+	if err := cliResources.Delete(ctx, obj); err != nil {
 		return fmt.Errorf("failed to delete resource %s: %w", obj.GetName(), err)
 	}
 
@@ -217,4 +230,160 @@ func updateSync(ctx context.Context, cfg *envconf.Config, obj paasapi.Resource, 
 	}
 
 	return waitForCondition(ctx, cfg, obj, gen, readyCondition)
+}
+
+func createPaasConfig(ctx context.Context, cfg *envconf.Config, paasConfig v1alpha2.PaasConfig) error {
+	// Create PaasConfig resource for testing
+	err := cfg.Client().Resources().Create(ctx, &paasConfig)
+	if err != nil {
+		return err
+	}
+
+	waitUntilPaasConfigExists := conditions.New(cfg.Client().Resources()).
+		ResourceMatch(&paasConfig, func(obj k8s.Object) bool {
+			return obj.(*v1alpha2.PaasConfig).Name == paasConfig.Name
+		})
+	return waitForDefaultOpts(ctx, waitUntilPaasConfigExists)
+}
+
+func retrieveBearerToken(
+	ctx context.Context,
+	cfg *envconf.Config,
+	namespace string,
+	secretName string,
+	key string,
+) error {
+	var tokenSecret = &corev1.Secret{}
+	if err := cfg.Client().Resources().Get(ctx, secretName, namespace, tokenSecret); err != nil {
+		return err
+	}
+	token, ok := tokenSecret.Data[key]
+	if !ok {
+		return fmt.Errorf("%s not in data for secret %s.%s", key, namespace, secretName)
+	}
+	pluginToken = string(token)
+	return nil
+}
+
+func registerSchemes(cfg *envconf.Config) error {
+	r, err := resources.New(cfg.Client().RESTConfig())
+	if err != nil {
+		return err
+	}
+	scheme := r.GetScheme()
+
+	for _, install := range []func(*runtime.Scheme) error{
+		v1alpha1.AddToScheme,
+		v1alpha2.AddToScheme,
+		quotav1.Install,
+		userv1.Install,
+	} {
+		install(scheme)
+	}
+
+	return nil
+}
+
+func startPortForward(
+	config *rest.Config,
+	namespace string,
+	serviceName string,
+	port int,
+) (int, func(), error) {
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return 0, nil, fmt.Errorf("failed to create clientset: %v", err)
+	}
+	podName, err := GetPodNameForService(clientset, namespace, serviceName)
+	if err != nil {
+		return 0, nil, err
+	}
+	req := clientset.CoreV1().RESTClient().Post().
+		Resource("pods").
+		Namespace(namespace).
+		Name(podName).
+		SubResource("portforward")
+
+	transport, upgrader, err := spdy.RoundTripperFor(config)
+	if err != nil {
+		return 0, nil, fmt.Errorf("error on round tripper: %w", err)
+	}
+
+	dialer := spdy.NewDialer(upgrader, &http.Client{Transport: transport}, "POST", req.URL())
+
+	// 2. Setup communication channels
+	stopChan := make(chan struct{}, 1)
+	readyChan := make(chan struct{})
+
+	// Buffer for error logs
+	out, errOut := new(bytes.Buffer), new(bytes.Buffer)
+
+	// Port 0 makes OS select an open port
+	ports := []string{fmt.Sprintf("0:%d", port)}
+
+	pf, err := portforward.New(dialer, ports, stopChan, readyChan, out, errOut)
+	if err != nil {
+		return 0, nil, fmt.Errorf("error while creating portforwarder: %w", err)
+	}
+
+	// 3. Start the forwarder in a goroutine
+	go func() {
+		if fwdErr := pf.ForwardPorts(); fwdErr != nil {
+			log.Fatalf("PortForward error: %v", fwdErr)
+		}
+	}()
+
+	// 4. Wait for tunnel to be ready
+	<-readyChan
+
+	// 5. Read port
+	forwardedPorts, err := pf.GetPorts()
+	if err != nil {
+		return 0, nil, fmt.Errorf("cannot retrieve port: %w", err)
+	}
+	localPort := int(forwardedPorts[0].Local)
+
+	// Return port and cleanup func (so you can close the port forward)
+	cleanup := func() {
+		close(stopChan)
+	}
+
+	return localPort, cleanup, nil
+}
+
+// GetPodNameForService finds a pod belonging to a service
+func GetPodNameForService(clientset *kubernetes.Clientset, namespace, serviceName string) (string, error) {
+	// 1. get Service
+	svc, err := clientset.CoreV1().Services(namespace).Get(context.TODO(), serviceName, metav1.GetOptions{})
+	if err != nil {
+		return "", fmt.Errorf("cannot find service: %w", err)
+	}
+
+	// Check for selectors
+	if len(svc.Spec.Selector) == 0 {
+		return "", fmt.Errorf("service %s has no selector", serviceName)
+	}
+
+	// 2. change map to string
+	set := labels.Set(svc.Spec.Selector)
+	listOptions := metav1.ListOptions{LabelSelector: set.AsSelector().String()}
+
+	// 3. find pods
+	pods, err := clientset.CoreV1().Pods(namespace).List(context.TODO(), listOptions)
+	if err != nil {
+		return "", fmt.Errorf("error searching for pods: %w", err)
+	}
+
+	if len(pods.Items) == 0 {
+		return "", fmt.Errorf("no pods for this service %s", serviceName)
+	}
+
+	// 4. Use first pod that is in a running state
+	for _, pod := range pods.Items {
+		if pod.Status.Phase == corev1.PodRunning {
+			return pod.Name, nil
+		}
+	}
+
+	return "", fmt.Errorf("no running pods for this service %s", serviceName)
 }
